requirements.txt
django
djangorestframework
instaloader
Mock